move humanizing fns with rest of number formatting
