+++ web-common/src/features/dashboards/humanize-numbers.ts
@@ -8,10 +8,6 @@ import {
   NumberKind,
   NumberParts,
 } from "@rilldata/web-common/lib/number-formatting/humanizer-types";
-import {
-  formatMsInterval,
-  formatMsToDuckDbIntervalString,
-} from "@rilldata/web-common/lib/number-formatting/strategies/intervals";
 import { PerRangeFormatter } from "@rilldata/web-common/lib/number-formatting/strategies/per-range";
 
 /**
@@ -52,46 +48,6 @@ export const formatPresetToNumberKind = (type: FormatPreset | string) => {
       return NumberKind.ANY;
   }
 };
-
-export function humanizeDataType(value: unknown, type: FormatPreset): string {
-  if (value === undefined || value === null) return "";
-  if (typeof value !== "number") return value.toString();
-
-  const numberKind = formatPresetToNumberKind(type);
-
-  let innerOptions: FormatterFactoryOptions;
-
-  if (type === FormatPreset.NONE) {
-    innerOptions = {
-      strategy: "none",
-      numberKind,
-      padWithInsignificantZeros: false,
-    };
-  } else if (type === FormatPreset.INTERVAL) {
-    return formatMsInterval(value);
-  } else {
-    innerOptions = {
-      strategy: "default",
-      numberKind,
-    };
-  }
-  return humanizedFormatterFactory([value], innerOptions).stringFormat(value);
-}
-
-/**
- * This function is intended to provide a lossless
- * humanized string representation of a number in cases
- * where a raw number will be meaningless to the user.
- */
-export function humanizeDataTypeExpanded(
-  value: unknown,
-  type: FormatPreset
-): string {
-  if (type === FormatPreset.INTERVAL) {
-    return formatMsToDuckDbIntervalString(value as number);
-  }
-  return value.toString();
-}
 
 /** This function is used primarily in the leaderboard and the detail tables. */
 export function humanizeDimTableValue(value: number, type: FormatPreset) {
